pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contextgate"
version = "0.1.0"
description = "Demonstration of retrieval-layer data boundaries for LLM applications."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ContextGate" }]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39"]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
    "ignore::DeprecationWarning:importlib",
]
